requirements.txt
absl-py           ==    0.11.0
addict              ==  2.2.1
argon2-cffi          == 20.1.0
astor                == 0.8.1
async-generator      == 1.10
attrs                == 20.3.0
backcall            ==  0.2.0
bleach               == 3.2.1
cached-property     ==  1.5.2
cachetools          ==  4.1.1
certifi             ==  2020.6.20
cffi            ==      1.14.3
chardet           ==    3.0.4
cycler              ==  0.10.0
Cython               == 0.29.21
decorator          ==   4.4.2
defusedxml        ==    0.6.0
descartes          ==   1.1.0
easydict          ==    1.9
einops            ==    0.3.0
entrypoints        ==   0.3
filterpy            ==  1.4.5
fire            ==      0.3.1
future           ==     0.18.2
gast              ==    0.2.2
google-pasta      ==    0.2.0
graphviz          ==    0.14.2
grpcio            ==    1.33.2
h5py               ==   3.0.0
idna               ==   2.10
imagecorruptions   ==   1.1.0
imageio            ==   2.9.0
importlib-metadata   == 2.0.0
ipykernel            == 5.3.4
ipython             ==  7.19.0
ipython-genutils    ==  0.2.0
ipywidgets          ==  7.5.1
jedi               ==   0.17.2
Jinja2             ==   2.11.2
joblib             ==   0.16.0
jsonschema         ==   3.2.0
jupyter             ==  1.0.0
jupyter-client    ==    6.1.7
jupyter-console    ==   6.2.0
jupyter-core       ==   4.6.3
jupyterlab-pygments ==  0.1.2
Keras-Applications  ==  1.0.8
Keras-Preprocessing  == 1.1.2
kiwisolver          ==  1.3.1
llvmlite           ==   0.33.0
Markdown           ==   3.3.3
MarkupSafe          ==  1.1.1
matplotlib          ==  3.3.2
mistune             ==  0.8.4
motmetrics          ==  1.1.3
nbclient            ==  0.5.1
nbconvert           ==  6.0.7
nbformat            ==  5.0.8
nest-asyncio         == 1.4.2
networkx            ==  2.5
notebook            ==  6.1.5
numba              ==   0.50.1
numpy              ==   1.18.5
nuscenes-devkit    ==   1.1.0
objgraph           ==   3.5.0
open3d-python      ==   0.7.0.0
opencv-contrib-python== 4.4.0.46
opencv-python       ==  4.3.0.38
opt-einsum          ==  3.3.0
packaging           ==  20.4
pandas              ==  1.1.4
pandocfilters       ==  1.4.3
parso               ==  0.7.1
pexpect             ==  4.8.0
pickleshare         ==  0.7.5
Pillow              ==  6.2.1
pip                 ==  20.2.4
prometheus-client   ==  0.8.0
prompt-toolkit      ==  3.0.8
protobuf            ==  3.13.0
ptyprocess         ==   0.6.0
pybind11           ==   2.5.0
pycocotools         ==  2.0.2
pycparser           ==  2.20
Pygments            ==  2.7.2
pyparsing           ==  2.4.7
pyquaternion        ==  0.9.9
pyrsistent           == 0.17.3
pytest-runner       ==  5.2
python-dateutil     ==  2.8.1
pytz                ==  2020.4
PyWavelets == 1.1.1
PyYAML        ==        5.3.1
pyzmq           ==      19.0.2
qtconsole         ==    4.7.7
QtPy                ==  1.9.0
requests             == 2.24.0
scikit-image       ==   0.17.2
scikit-learn      ==    0.22.2.post1
scipy               ==  1.4.1
Send2Trash         ==   1.5.0
setuptools          ==  49.2.1
Shapely             ==  1.7.1
siren-pytorch      ==   0.0.9
six                ==   1.15.0
tensorboard        ==   1.15.0
tensorflow-estimator == 1.15.1
tensorflow-gpu      ==  1.15.4
termcolor          ==   1.1.0
terminado          ==   0.9.1
terminaltables     ==   3.1.0
testpath           ==   0.4.4
tifffile    ==          2020.10.1
torch         ==        1.6.0
torchvision     ==      0.7.0
tornado           ==    6.1
tqdm                ==  4.48.2
traitlets             ==5.0.5
transformations==       2020.1.1
urllib3          ==     1.25.11
vtk                ==   9.0.1
wcwidth              == 0.2.5
webencodings==          0.5.1
Werkzeug      ==        1.0.1
wheel           ==      0.35.1
widgetsnbextension==    3.5.1
wrapt               ==  1.12.1
zipp                  ==3.4.0